+++ packages/react-components/react-button/src/components/CompoundButton/CompoundButton.tsx
@@ -4,6 +4,7 @@ import { useCompoundButton_unstable } from './useCompoundButton';
 import { useCompoundButtonStyles_unstable } from './useCompoundButtonStyles';
 import type { ForwardRefComponent } from '@fluentui/react-utilities';
 import type { CompoundButtonProps } from './CompoundButton.types';
+import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * CompoundButtons are buttons that can have secondary content that adds extra information to the user.
@@ -12,6 +13,9 @@ export const CompoundButton: ForwardRefComponent<CompoundButtonProps> = React.fo
   const state = useCompoundButton_unstable(props, ref);
 
   useCompoundButtonStyles_unstable(state);
+
+  const { useCompoundButtonStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
+  useCustomStyles(state);
 
   return renderCompoundButton_unstable(state);
   // Casting is required due to lack of distributive union to support unions on @types/react