+++ packages/react-components/react-button/src/components/SplitButton/SplitButton.tsx
@@ -4,6 +4,7 @@ import { useSplitButton_unstable } from './useSplitButton';
 import { useSplitButtonStyles_unstable } from './useSplitButtonStyles';
 import type { SplitButtonProps } from './SplitButton.types';
 import { ForwardRefComponent } from '@fluentui/react-utilities';
+import { useCustomStyleHooks_unstable } from '@fluentui/react-shared-contexts';
 
 /**
  * SplitButtons are a grouping of two interactive surfaces where interacting with the first one triggers a primary
@@ -13,6 +14,9 @@ export const SplitButton: ForwardRefComponent<SplitButtonProps> = React.forwardR
   const state = useSplitButton_unstable(props, ref);
 
   useSplitButtonStyles_unstable(state);
+
+  const { useSplitButtonStyles_unstable: useCustomStyles } = useCustomStyleHooks_unstable();
+  useCustomStyles(state);
 
   return renderSplitButton_unstable(state);
   // Casting is required due to lack of distributive union to support unions on @types/react