+++ packages/react-components/react-provider/src/components/FluentProvider/useFluentProvider.ts
@@ -3,8 +3,13 @@ import {
   ThemeContext_unstable as ThemeContext,
   useFluent_unstable as useFluent,
   useOverrides_unstable as useOverrides,
+  useCustomStyleHooks_unstable as useCustomStyleHooks,
 } from '@fluentui/react-shared-contexts';
-import type { ThemeContextValue_unstable as ThemeContextValue } from '@fluentui/react-shared-contexts';
+import type {
+  CustomStyleHooksContextValue_unstable as CustomStyleHooksContextValue,
+  ThemeContextValue_unstable as ThemeContextValue,
+} from '@fluentui/react-shared-contexts';
+
 import { getNativeElementProps, useMergedRefs } from '@fluentui/react-utilities';
 import * as React from 'react';
 import { useFluentProviderThemeStyleTag } from './useFluentProviderThemeStyleTag';
@@ -26,6 +31,7 @@ export const useFluentProvider_unstable = (
   const parentContext = useFluent();
   const parentTheme = useTheme();
   const parentOverrides = useOverrides();
+  const parentCustomStyleHooks = useCustomStyleHooks();
 
   /**
    * TODO: add merge functions to "dir" merge,
@@ -34,6 +40,8 @@ export const useFluentProvider_unstable = (
    */
   const {
     applyStylesToPortals = true,
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    customStyleHooks_unstable,
     dir = parentContext.dir,
     targetDocument = parentContext.targetDocument,
     theme,
@@ -42,6 +50,12 @@ export const useFluentProvider_unstable = (
   const mergedTheme = shallowMerge(parentTheme, theme);
 
   const mergedOverrides = shallowMerge(parentOverrides, overrides);
+
+  // parentCustomStyleHooks will not be a partial
+  const mergedCustomStyleHooks = shallowMerge(
+    parentCustomStyleHooks,
+    customStyleHooks_unstable,
+  ) as CustomStyleHooksContextValue;
 
   React.useEffect(() => {
     if (process.env.NODE_ENV !== 'production' && mergedTheme === undefined) {
@@ -57,6 +71,8 @@ export const useFluentProvider_unstable = (
 
   return {
     applyStylesToPortals,
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    customStyleHooks_unstable: mergedCustomStyleHooks,
     dir,
     targetDocument,
     theme: mergedTheme,