+++ packages/react/src/components/Button/BaseButton.tsx
@@ -699,14 +699,8 @@ export class BaseButton extends React.Component<IBaseButtonProps, IBaseButtonSta
     classNames: ISplitButtonClassNames | undefined,
     keytipAttributes: any,
   ): JSX.Element {
-    const {
-      allowDisabledFocus,
-      checked,
-      disabled,
-      splitButtonMenuProps,
-      splitButtonAriaLabel,
-      primaryDisabled,
-    } = this.props;
+    const { allowDisabledFocus, checked, disabled, splitButtonMenuProps, splitButtonAriaLabel, primaryDisabled } =
+      this.props;
     const { menuHidden } = this.state;
     let menuIconProps = this.props.menuIconProps;